requirements.txt
groq
requests
python-dotenv
google-api-python-client
google-auth-oauthlib
edge-tts